pyproject.toml
[project]
name = "agent-ochestration-sk-autogen"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "semantic-kernel (>=1.35.0,<2.0.0)",
    "autogen-agentchat (>=0.7.1,<0.8.0)",
    "autogen-ext[azure,openai] (>=0.7.1,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
